DELIA-21705 : Change the variable type to long long to support big numbers.
Reason for change: we used 'unsigned long' to report/return the capacity n freespace of the memory devices but noticed that this variable is 4 bytes size but the XG platforms has big sizes n the value must be reported in 8 byte length variable to represent properly. Fixed it now. This needs a API change, the XRE/ServiceManager, HostIf, SNMP will be modified to fix this.
Test Procedure: Please verify this in XG platforms
Risks: Medium
Signed-off-by: Karunakaran A <Karunakaran_Amirthalingam@cable.comcast.com>

Change-Id: I3bc75036201499b85f0e38ef98a020c406b8d9ee